fix 42Oauth -> logout -> 42Oauthの順で動作しなかったので修正
