+++ django/backend/accounts/templates/accounts/login-signup.html
@@ -37,10 +37,10 @@
     //import { LoginEvent } from '../../../accounts/js/login.js';
     //import { SignupEvent } from '../../../accounts/js/signup.js';
     try {
-      const LoginEvent = new Event('LoginEvent');
-      const SignupEvent = new Event('SignupEvent');
-      document.dispatchEvent(LoginEvent);
-      document.dispatchEvent(SignupEvent);
+      //const LoginEvent = new Event('LoginEvent');
+      //const SignupEvent = new Event('SignupEvent');
+      //document.dispatchEvent(LoginEvent);
+      //document.dispatchEvent(SignupEvent);
 
       const login_link = document.getElementById('tab-login');
       const signup_link = document.getElementById('tab-signup');
