+++ django/backend/accounts/views.py
@@ -4,6 +4,8 @@ from django.views.generic import CreateView
 from django.contrib.auth import login, authenticate
 from django.contrib.auth.views import LoginView, LogoutView
 
+# from django.views.decorators.csrf import csrf_exempt
+
 # from django.contrib.auth.forms import AuthenticationForm
 # from django.template.exceptions import TemplateDoesNotExist
 from django.utils.decorators import method_decorator
