+++ django/backend/spa/templates/spa/nav.html
@@ -61,7 +61,6 @@
                     </button>
                   </form>
                 </div>
-
               </div>
             </div>
           </li>
