+++ django/frontend/src/index.js
@@ -3,14 +3,10 @@ import { Routes } from './spa/js/routing/routes.js';
 import { navigateTo, updatePage } from './spa/js/routing/routing.js';
 import { changingLanguage } from './spa/js/utility/lang.js';
 import { getUrl } from './spa/js/utility/url.js';
-//import { fetchAsForm, fetchData } from './spa/js/utility/fetch.js';
 import { fetchAsForm } from './spa/js/utility/fetch.js';
-import { getUrlWithLang } from './spa/js/utility/url.js';
-//import { getUrlWithLang } from '../utility/url.js';
-import fetchData from './spa/js/utility/fetch.js';
-//import { Pills } from 'bootstrap.bundle.min.js';
-//import Cookies from 'js-cookie';
-//import { Tooltip, Toast, Popover } from 'bootstrap';
+//import { getUrlWithLang } from './spa/js/utility/url.js';
+//import { getUrlWithLang } from './spa/js/utility/url.js';
+//import fetchData from './spa/js/utility/fetch.js';
 import 'bootstrap';
 import './accounts/js/two_fa.js';
 import './accounts/js/login.js';
@@ -18,7 +14,7 @@ import './accounts/js/signup.js';
 
 import './spa/scss/spa.scss';
 import './main.scss';
-import { logout } from './spa/js/utility/user.js';
+import { loadNav } from './spa/js/utility/user.js';
 
 console.log('test No.1');
 
@@ -32,22 +28,15 @@ const getDisplayedURI = (pathname) => {
   return getUrl(path);
 };
 
-const loadNav = async () => {
-  try {
-    const nav_uri = getUrlWithLang('spa/nav');
-    const nav_html = await fetchData(nav_uri);
-    document.querySelector('#nav').innerHTML = nav_html;
-
-    const logout_button = document.getElementById('nav-logout-button');
-    logout_button.addEventListener('click', async () => {
-      await logout();
-    });
-  } catch (error) {
-    console.warning('ignore:' + error);
-  }
-};
-
 document.addEventListener('DOMContentLoaded', async () => {
+  console.log('load No.1');
+
+  /*
+  const LoginEvent = new Event('LoginEvent');
+  const SignupEvent = new Event('SignupEvent');
+  document.dispatchEvent(LoginEvent);
+  document.dispatchEvent(SignupEvent);
+  */
   loadNav();
   /*
   console.log('nav test No.2');
