+++ django/frontend/src/spa/js/utility/user.js
@@ -1,5 +1,7 @@
 import { fetchAsForm } from './fetch.js';
 import { router } from '../routing/routing.js';
+import { getUrlWithLang } from './url.js';
+import fetchData from './fetch.js';
 
 export function isLogined() {
   //todo
@@ -11,6 +13,27 @@ export function isLogined() {
 
 export const reload = async () => {
   await router();
+  await loadNav();
+};
+
+export const loadNav = async () => {
+  try {
+    const nav_uri = getUrlWithLang('spa/nav');
+    const nav_html = await fetchData(nav_uri);
+    document.querySelector('#nav').innerHTML = nav_html;
+
+    const logout_button = document.getElementById('nav-logout-button');
+    logout_button.addEventListener('click', async () => {
+      await logout();
+    });
+  } catch (error) {
+    console.warning('ignore:' + error);
+  }
+};
+
+export const handlePostLogin = async () => {
+  await reload();
+  document.getElementById('nav').hidden = false;
 };
 
 export const logout = async () => {
@@ -18,11 +41,7 @@ export const logout = async () => {
   const formData = new FormData(form);
   const response = await fetchAsForm(form, formData);
   if (response.status == 200) {
-    document.getElementById('nav').hidden = true;
     await reload();
+    document.getElementById('nav').hidden = true;
   }
 };
-
-//const reload = async () => {
-//      router();
-//};
